backend/requirements.txt
fastapi
uvicorn[standard]
websockets
pydantic
httpx
python-dotenv
supabase
